wip, shoehorning in similar issues
